Fix search error & add collapse to characterList
