+++ pages/CharacterList/TestData.js
@@ -1,0 +1,100 @@
+const TestData = [
+  {
+    id: "1",
+    name: "6migard",
+    rarity: "6",
+    main_attribute: "forest",
+    sub_attribute: "forest",
+    class: "sniper",
+  },
+  {
+    id: "2",
+    name: "5migard",
+    rarity: "5",
+    main_attribute: "forest",
+    sub_attribute: "forest",
+    class: "sniper",
+  },
+  {
+    id: "3",
+    name: "4migard",
+    rarity: "4",
+    main_attribute: "forest",
+    sub_attribute: "forest",
+    class: "sniper",
+  },
+  {
+    id: "4",
+    name: "6karen",
+    rarity: "6",
+    main_attribute: "water",
+    sub_attribute: "water",
+    class: "changer",
+  },
+  {
+    id: "5",
+    name: "3karen",
+    rarity: "3",
+    main_attribute: "water",
+    sub_attribute: "water",
+    class: "changer", 
+  },
+  {
+    id: "6",
+    name: "4karen",
+    rarity: "4",
+    main_attribute: "water",
+    sub_attribute: "water",
+    class: "changer",
+  },
+  {
+    id: "12",
+    name: "6migard",
+    rarity: "6",
+    main_attribute: "forest",
+    sub_attribute: "forest",
+    class: "sniper",
+  },
+  {
+    id: "7",
+    name: "5migard",
+    rarity: "5",
+    main_attribute: "forest",
+    sub_attribute: "forest",
+    class: "sniper",
+  },
+  {
+    id: "8",
+    name: "4migard",
+    rarity: "4",
+    main_attribute: "forest",
+    sub_attribute: "forest",
+    class: "sniper",
+  },
+  {
+    id: "9",
+    name: "6karen",
+    rarity: "6",
+    main_attribute: "water",
+    sub_attribute: "water",
+    class: "changer",
+  },
+  {
+    id: "10",
+    name: "3karen",
+    rarity: "3",
+    main_attribute: "water",
+    sub_attribute: "water",
+    class: "changer", 
+  },
+  {
+    id: "11",
+    name: "4karen",
+    rarity: "4",
+    main_attribute: "water",
+    sub_attribute: "water",
+    class: "changer",
+  }
+]
+
+export default TestData;
